requirements.txt
selenium
pytesseract
uvicorn
fastapi
pillow
pony
webdriver_manager
keyboard
pyperclip
pyvirtualdisplay
python-xlib